Add files via upload
In this mini-project, a network has been designed for International Apollo University where
six campuses will have an internal connection. The whole network design was developed
using the Cisco Packet Tracer network simulation program. In design, all the campuses hold a
specific network. The main campus will have a network only for the servers. Any user can
also make a wireless connection for certain parts of information. There is also multiple
website and servers for this university. The DHCP server will provide all the important
information to all hosts automatically. Any host of one campus will be able to communicate
with other hosts of all campuses. The university must have a DNS and WEB server to show
the website of the university.